[BEAM-1871] Move GCP specific serialization CloudObject and supporting translation code to Dataflow runner module
