+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/util/CloudObjectTranslator.java
@@ -18,8 +18,6 @@
 
 package org.apache.beam.runners.dataflow.util;
 
-import org.apache.beam.sdk.util.CloudObject;
-
 /**
  * A translator that takes an object and creates a {@link CloudObject} which can be converted back
  * to the original object.
